% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{Combine_Reduce}
\alias{Combine_Reduce}
\title{List Combining and Reducing}
\usage{
Combine_Reduce(combine, multicombine, result)
}
\arguments{
\item{combine}{function that is used to process the tasks results as they generated.}

\item{multicombine}{logical flag indicating whether the \code{combine} function can accept more than two arguments.
If TRUE, then the \code{combine} function will directely combine the \code{result} to final output.
If FALSE, then the \code{combine} function will always be called with two arguments and reduce to final output.}

\item{result}{list to be processed.}
}
\value{
combined or reduced result
}
\description{
List Combining and Reducing
}
\keyword{internal}
