% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForEach.R
\name{ForEach}
\alias{ForEach}
\alias{\%run\%}
\title{ForEach}
\usage{
ForEach(..., .combine = "list", .multicombine = TRUE, cores = 1L)

iter \%run\% expr
}
\arguments{
\item{...}{one argument that control how expression is evaluated.
Named arguments specify the name and values of variables to be defined in the evaluation environment.}

\item{.combine}{function that is used to process the tasks results as they generated.}

\item{.multicombine}{logical flag indicating whether the \code{combine} function can accept more than two arguments.
If TRUE, then the \code{combine} function will directely combine the \code{result} to final output.
If FALSE, then the \code{combine} function will always be called with two arguments and reduce to final output.}

\item{cores}{The number of cores to use.}

\item{iter}{ForEach object.}

\item{expr}{the R expression to evaluate.}
}
\description{
Binary operator %run% that operates on a ForEach object and an R expression.
}
\examples{
ForEach(i = 1:5, .combine = "c", .multicombine = TRUE) \%run\% {
  rnorm(1)
}
ForEach(i = 1:5, .combine = "union", .multicombine = FALSE) \%run\% {
  round(runif(1), digits = 1)
}
}
