% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process.R
\name{Process}
\alias{Process}
\alias{Collect}
\title{Process}
\usage{
Process(expr)

Collect(process, .combine = "list", .multicombine = TRUE)
}
\arguments{
\item{expr}{the R expression to evaluate.}

\item{process}{a list of processes to be collected.}

\item{.combine}{function that is used to process the tasks results as they generated.}

\item{.multicombine}{logical flag indicating whether the \code{combine} function can accept more than two arguments.
If TRUE, then the \code{combine} function will directely combine the \code{result} to final output.
If FALSE, then the \code{combine} function will always be called with two arguments and reduce to final output.}
}
\description{
\code{Process} starts a parallel R process which evaluates the given expression.

\code{Collect} collects results from one or more parallel processes.
}
\examples{
task1 <- Process(rnorm(1))
task2 <- Process(runif(1))
Collect(list(task1, task2), .combine = "c")
}
