% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Promise.R
\name{Promise}
\alias{Promise}
\alias{Execute}
\title{Promise}
\usage{
Promise(expr)

Execute(promise, .combine = "list", .multicombine = TRUE)
}
\arguments{
\item{expr}{the R expression to evaluate.}

\item{promise}{a list of promises to be executed}

\item{.combine}{function that is used to process the tasks results as they generated.}

\item{.multicombine}{logical flag indicating whether the \code{combine} function can accept more than two arguments.
If TRUE, then the \code{combine} function will directely combine the \code{result} to final output.
If FALSE, then the \code{combine} function will always be called with two arguments and reduce to final output.}
}
\description{
\code{Promise} creates a promise that evaluates an R expression.

\code{Execute} evaluates promises and get the results.
}
\examples{
task1 <- Promise(rnorm(1))
task2 <- Promise(runif(1))
Execute(list(task1, task2), .combine = "c")
}
